requirements.txt
pandas== 1.4.2
scikit-learn== 1.2.2
Flask== 2.1.2
pymongo== 4.3.3
python-decouple== 3.8